requirements.txt
# PFA Planning System Dependencies
PyYAML>=6.0
requests>=2.25.1
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.2
python-dotenv>=0.19.0